requirements.txt
Flask==2.3.3
openai==1.3.0
groq==0.4.0
python-dotenv==1.0.0
selenium==4.15.0
pytest==7.4.3
jinja2==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.0 